Add Role service

+++ bookMyHotel/bookMyHotel/src/main/java/com/project/bookMyHotel/service/UserService.java
@@ -41,7 +41,10 @@ public class UserService implements IUserService{
     @Transactional
     @Override
     public void deleteUser(String email) {
-        userRepository.deleteByEmail(email);
+        User theUser = getUser(email);
+        if(theUser != null){
+            userRepository.deleteByEmail(email);
+        }
     }
 
     @Override
